# openapi.json
{"openapi":"3.1.0","info":{"title":"API da Aula 2","description":"\n    API desenvolvida durante a aula 2 (bloco c), contendo endpoints de exemplo e soma\"\n\n    - /teste: retorna uma mensagem de sucesso\n    - /soma/numero1/numero2: recebe dois números e retorna a soma\n","termsOfService":"http://example.com/terms/","contact":{"name":"Samantha Adiely Alecrim","url":"http://github.com/Adiely/","email":"samanthaalecrim.biotec@gmail.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"0.1"},"paths":{"/teste":{"get":{"tags":["Teste"],"summary":"Retorna mensagem de teste","description":"Retorna uma mensagem de exemplo para testar e verificar se deu certo","operationId":"hello_world_teste_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/soma/{numero1}/{numero2}/{api_token}":{"post":{"tags":["Operações matemáticas simples enum"],"summary":"Soma","operationId":"soma_soma__numero1___numero2___api_token__post","parameters":[{"name":"numero1","in":"path","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"path","required":true,"schema":{"type":"integer","title":"Numero2"}},{"name":"api_token","in":"path","required":true,"schema":{"type":"integer","title":"Api Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/soma/v2":{"post":{"tags":["Operações matemáticas simples enum"],"summary":"Soma Formato2","operationId":"soma_formato2_soma_v2_post","parameters":[{"name":"numero1","in":"query","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"query","required":true,"schema":{"type":"integer","title":"Numero2"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/soma/v3":{"post":{"tags":["Operações matemáticas simples enum"],"summary":"Soma Formato3","operationId":"soma_formato3_soma_v3_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Numeros"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resultado"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/divisao/{numero1}/{numero2}":{"post":{"tags":["Operações matemáticas simples enum"],"summary":"Divisao","operationId":"divisao_divisao__numero1___numero2__post","parameters":[{"name":"numero1","in":"path","required":true,"schema":{"type":"integer","title":"Numero1"}},{"name":"numero2","in":"path","required":true,"schema":{"type":"integer","title":"Numero2"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Numeros":{"properties":{"numero1":{"type":"integer","title":"Numero1"},"numero2":{"type":"integer","title":"Numero2"},"numero3":{"type":"integer","title":"Numero3","default":0}},"type":"object","required":["numero1","numero2"],"title":"Numeros"},"Resultado":{"properties":{"resultado":{"type":"integer","title":"Resultado"}},"type":"object","required":["resultado"],"title":"Resultado"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}